♻️ Refactor CORS middleware in index.js
