+++ app/index.js
@@ -3,32 +3,18 @@ const connectToMongoDB = require("./db/setupDb");
 
 const express = require("express");
 const logger = require("./middleware/loggerMiddleware");
-const cors = require("cors");
+const corsMiddleware = require("./middleware/corsMiddleware");
+const verifyToken = require("./middleware/verifyTokenMiddleware");
 
 const usersRoutes = require("./routes/v1/usersRoutes");
 const authRoutes = require("./routes/authRoutes");
-const verifyToken = require("./middleware/verifyTokenMiddleware");
 
 const app = express();
 
-const corsOptions = {
-  origin: process.env.CORS_ORIGIN.split(","),
-  optionsSuccessStatus: 200,
-  methods: "GET, POST, PATCH, DELETE",
-  allowedHeaders: [
-    "Content-Type",
-    "Authorization",
-    "Origin",
-    "X-Requested-With",
-    "Username",
-  ],
-  credentials: true,
-};
-
 app.use(express.json());
 app.use(logger);
-app.use("/api/v1/users", cors(corsOptions), verifyToken, usersRoutes);
-app.use("/auth/", cors(corsOptions), authRoutes);
+app.use("/api/v1/users", corsMiddleware, verifyToken, usersRoutes);
+app.use("/auth/", corsMiddleware, authRoutes);
 
 const port = process.env.LISTENING_PORT;
